Taking another attempt to use rep-color more prominently on cards and product pages.
Previous attempts were not successful because the rep color was mixed with background as is - resulting it a very subtle effect depending on the rep color luminosity and theme (dark/light). Obviously brighter colors worked better in dark themes and darker in light. That also resulted in many cards not getting noticeable treatment.

This attempt is different in those key ways:
- Before mixing with highlight, rep color is mixed with foreground that guarantees higher contrast ratio (foreground is a contrast color to highlight in every theme)
- Cards are getting a border based on rep color mixed with foreground, that is also expected to be noticeable against the card background (lighter in dark theme, darker in light theme). What's most important is that this color is further away from the background color compared to card background and cards stand out more against the background
- Mix ratios have been tweaked with all of the above in mind resulting in much more vibrant, but still decently contrasting color combinations